+++ weechat_script_lint/main.py
@@ -20,7 +20,7 @@
 
 """Static analysis tool for WeeChat scripts."""
 
-from typing import Dict, Generator, Tuple
+from typing import Dict, Generator, List, Tuple
 
 import argparse
 import pathlib
@@ -97,7 +97,10 @@ def get_parser() -> argparse.ArgumentParser:
         ),
     )
     parser.add_argument(
-        "-q", "--quiet", action="store_true", help="do not display any message"
+        "-q",
+        "--quiet",
+        action="store_true",
+        help="do not display any message",
     )
     parser.add_argument(
         "-r",
@@ -110,6 +113,15 @@ def get_parser() -> argparse.ArgumentParser:
         "--strict",
         action="store_true",
         help="count warnings as errors in the returned code",
+    )
+    parser.add_argument(
+        "-S",
+        "--score",
+        action="store_true",
+        help=(
+            "display scores by script, grouped by score, "
+            "do not display report and return code"
+        ),
     )
     parser.add_argument(
         "-v", "--verbose", action="store_true", help="verbose output"
@@ -125,29 +137,26 @@ def get_parser() -> argparse.ArgumentParser:
 
 
 def get_scripts(
-    path: pathlib.Path, recursive: bool
+    path: pathlib.Path, args, ignored_files: List[str]
 ) -> Generator[pathlib.Path, None, None]:
     """
     Return the list of scripts in a path.
 
-    :param path: path
-    :param recursive: recursively list scripts in sub-directories
+    :param path: path (directory or file)
+    :param argparse.Namespace args: command-line arguments
     :return: list of scripts
     """
     if path.is_dir():
         for path2 in path.iterdir():
-            # ignore hidden files/directories
-            if path2.name.startswith("."):
-                continue
-            if path2.is_file():
-                if path2.suffix in SUPPORTED_SUFFIXES:
-                    yield path2
-            elif recursive and path2.is_dir():
-                yield from get_scripts(path2, recursive)
+            yield from get_scripts(path2, args, ignored_files)
     elif not path.is_file():
         sys.exit(f"FATAL: not a directory/file: {path}")
-    elif path.suffix in SUPPORTED_SUFFIXES:
-        yield path
+    elif not path.name.startswith(".") and path.suffix in SUPPORTED_SUFFIXES:
+        if path.name in ignored_files:
+            if not args.quiet and args.verbose:
+                print(f"{path}: file ignored")
+        else:
+            yield path
 
 
 def print_report(
@@ -185,7 +194,64 @@ def print_report(
         )
 
 
-def check_scripts(args) -> int:
+def get_string_score(score: int, use_colors: bool = True) -> str:
+    """
+    Get string with score.
+
+    :param score: script score (between 0 and 100)
+    :param use_colors: True to use colors in output
+    """
+    colorize = color if use_colors else lambda x, y: x
+    if score < 50:
+        status_color = "bold,red"
+    elif score < 80:
+        status_color = "bold,yellow"
+    elif score < 100:
+        status_color = "bold,cyan"
+    else:
+        status_color = "bold,green"
+    return colorize(f"{score} / 100", status_color)
+
+
+def print_scripts_by_score(
+    scores: Dict[pathlib.Path, int], use_colors: bool = True
+) -> None:
+    """
+    Print list of scripts grouped by score.
+
+    :param scores: scores
+    :param use_colors: True to use colors in output
+    """
+    scripts_by_score: Dict[int, List[pathlib.Path]] = {}
+    for path, score in scores.items():
+        scripts_by_score.setdefault(score, []).append(path)
+    for score in sorted(scripts_by_score, reverse=True):
+        sorted_paths = sorted(scripts_by_score[score])
+        count_scripts = len(sorted_paths)
+        paths = "\n".join([f"  {path}" for path in sorted_paths])
+        print(
+            f"{count_scripts} scripts "
+            f"with score {get_string_score(score, use_colors)}:\n"
+            f"{paths}"
+        )
+
+
+def print_scores(
+    scores: Dict[pathlib.Path, int], use_colors: bool = True
+) -> None:
+    """
+    Print scores for all checked scripts.
+
+    :param name: script name
+    :param score: script score (between 0 and 100)
+    :param use_colors: True to use colors in output
+    """
+    for path, score in scores.items():
+        str_score = get_string_score(score, use_colors)
+        print(f"{path}: score = {str_score}")
+
+
+def check_scripts(args) -> Tuple[int, int]:
     """
     Check scripts.
 
@@ -199,15 +265,11 @@ def check_scripts(args) -> int:
     }
     num_scripts = 0
     num_scripts_with_issues = 0
+    scores: Dict[pathlib.Path, int] = {}
     ignored_files = (args.ignore_files or "").split(",")
     for path in args.path:
-        scripts = get_scripts(path, args.recursive)
+        scripts = get_scripts(path, args, ignored_files)
         for path_script in scripts:
-            # ignored file?
-            if path_script.name in ignored_files:
-                if not args.quiet and args.verbose:
-                    print(f"{path_script}: file ignored")
-                continue
             # check script
             num_scripts += 1
             script = WeechatScript(
@@ -218,31 +280,34 @@ def check_scripts(args) -> int:
             )
             script.check()
             report = script.get_report(args.name_only)
+            scores[path_script] = script.score
             if report:
                 num_scripts_with_issues += 1
-                if report and not args.quiet:
+                if report and not args.quiet and not args.score:
                     print(report)
             # add errors/warnings/info found
             for counter in script.count:
                 count[counter] += script.count[counter]
-    if not args.quiet and not args.name_only:
+    if not args.quiet and args.score:
+        print_scripts_by_score(scores, use_colors=not args.no_colors)
+    if not args.quiet and not args.name_only and not args.score:
+        print_scores(scores, use_colors=not args.no_colors)
+    if not args.quiet and not args.name_only and not args.score:
         print_report(
             num_scripts,
             num_scripts_with_issues,
             count,
             use_colors=not args.no_colors,
         )
-    if args.strict:
-        return count["error"] + count["warning"]
-    return count["error"]
+    return (count["error"], count["warning"])
 
 
 def main() -> None:
     """Main function."""
     args = get_parser().parse_args()
-    errors = check_scripts(args)
-    ret_code = min(255, errors)
-    if not args.quiet and not args.name_only:
+    errors, warnings = check_scripts(args)
+    ret_code = min(255, errors + warnings if args.strict else errors)
+    if not args.quiet and not args.name_only and not args.score:
         print(f"Exiting with code {ret_code}")
     sys.exit(ret_code)